Refactor transport loading and add comprehensive tests
- Updated import paths for Transport and TransportLoader classes.
- Changed the method of loading transports from resources to use TransportLoader.
- Introduced a new test class TransportLoaderTest with extensive unit tests for the TransportLoader functionality.
- Added tests for various scenarios including basic transport parsing, handling of empty and comment lines, multiple transports from the same origin, skill and item requirements, quest requirements, varbit and varplayer requirements, and more.
- Ensured that malformed data is handled gracefully and that transports are not created for undefined origins or destinations.
- Implemented tests for filtering based on radius thresholds and ensuring that only valid transports are created.
