+++ src/main/java/shortestpath/PathMapOverlay.java
@@ -16,6 +16,8 @@ import net.runelite.client.ui.overlay.Overlay;
 import net.runelite.client.ui.overlay.OverlayLayer;
 import net.runelite.client.ui.overlay.OverlayPosition;
 import shortestpath.pathfinder.CollisionMap;
+import shortestpath.transport.Transport;
+import shortestpath.transport.TransportType;
 
 public class PathMapOverlay extends Overlay {
     private final Client client;
