+++ src/main/java/shortestpath/pathfinder/PathfinderConfig.java
@@ -4,6 +4,30 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+
+import static shortestpath.transport.TransportType.AGILITY_SHORTCUT;
+import static shortestpath.transport.TransportType.BOAT;
+import static shortestpath.transport.TransportType.CANOE;
+import static shortestpath.transport.TransportType.CHARTER_SHIP;
+import static shortestpath.transport.TransportType.FAIRY_RING;
+import static shortestpath.transport.TransportType.GNOME_GLIDER;
+import static shortestpath.transport.TransportType.GRAPPLE_SHORTCUT;
+import static shortestpath.transport.TransportType.HOT_AIR_BALLOON;
+import static shortestpath.transport.TransportType.MAGIC_CARPET;
+import static shortestpath.transport.TransportType.MAGIC_MUSHTREE;
+import static shortestpath.transport.TransportType.MINECART;
+import static shortestpath.transport.TransportType.QUETZAL;
+import static shortestpath.transport.TransportType.SEASONAL_TRANSPORTS;
+import static shortestpath.transport.TransportType.SHIP;
+import static shortestpath.transport.TransportType.SPIRIT_TREE;
+import static shortestpath.transport.TransportType.TELEPORTATION_BOX;
+import static shortestpath.transport.TransportType.TELEPORTATION_ITEM;
+import static shortestpath.transport.TransportType.TELEPORTATION_LEVER;
+import static shortestpath.transport.TransportType.TELEPORTATION_MINIGAME;
+import static shortestpath.transport.TransportType.TELEPORTATION_PORTAL;
+import static shortestpath.transport.TransportType.TELEPORTATION_SPELL;
+import static shortestpath.transport.TransportType.WILDERNESS_OBELISK;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Set;
@@ -28,34 +52,13 @@ import shortestpath.ShortestPathPlugin;
 import shortestpath.Destination;
 import shortestpath.ItemVariations;
 import shortestpath.PrimitiveIntHashMap;
-import shortestpath.Transport;
-import shortestpath.TransportItems;
-import shortestpath.TransportType;
-import shortestpath.TransportVarbit;
-import shortestpath.TransportVarPlayer;
 import shortestpath.WorldPointUtil;
-import static shortestpath.TransportType.AGILITY_SHORTCUT;
-import static shortestpath.TransportType.GRAPPLE_SHORTCUT;
-import static shortestpath.TransportType.BOAT;
-import static shortestpath.TransportType.CANOE;
-import static shortestpath.TransportType.CHARTER_SHIP;
-import static shortestpath.TransportType.SHIP;
-import static shortestpath.TransportType.FAIRY_RING;
-import static shortestpath.TransportType.GNOME_GLIDER;
-import static shortestpath.TransportType.HOT_AIR_BALLOON;
-import static shortestpath.TransportType.MAGIC_CARPET;
-import static shortestpath.TransportType.MAGIC_MUSHTREE;
-import static shortestpath.TransportType.MINECART;
-import static shortestpath.TransportType.QUETZAL;
-import static shortestpath.TransportType.SEASONAL_TRANSPORTS;
-import static shortestpath.TransportType.SPIRIT_TREE;
-import static shortestpath.TransportType.TELEPORTATION_BOX;
-import static shortestpath.TransportType.TELEPORTATION_LEVER;
-import static shortestpath.TransportType.TELEPORTATION_MINIGAME;
-import static shortestpath.TransportType.TELEPORTATION_PORTAL;
-import static shortestpath.TransportType.TELEPORTATION_ITEM;
-import static shortestpath.TransportType.TELEPORTATION_SPELL;
-import static shortestpath.TransportType.WILDERNESS_OBELISK;
+import shortestpath.transport.Transport;
+import shortestpath.transport.TransportItems;
+import shortestpath.transport.TransportLoader;
+import shortestpath.transport.TransportType;
+import shortestpath.transport.TransportVarPlayer;
+import shortestpath.transport.TransportVarbit;
 
 public class PathfinderConfig {
     private static final List<Integer> RUNE_POUCHES = Arrays.asList(
@@ -138,7 +141,7 @@ public class PathfinderConfig {
         this.config = config;
         this.mapData = SplitFlagMap.fromResources();
         this.map = ThreadLocal.withInitial(() -> new CollisionMap(mapData));
-        this.allTransports = Transport.loadAllFromResources();
+        this.allTransports = TransportLoader.loadAllFromResources();
         this.usableTeleports = new HashSet<>(allTransports.size() / 20);
         this.transports = new HashMap<>(allTransports.size() / 2);
         this.transportsPacked = new PrimitiveIntHashMap<>(allTransports.size() / 2);
