+++ src/main/java/shortestpath/transport/Transport.java
@@ -1,15 +1,15 @@
-package shortestpath;
+package shortestpath.transport;
 
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
-import java.util.Scanner;
 import java.util.Set;
 import lombok.Getter;
 import net.runelite.api.Quest;
 import net.runelite.api.Skill;
+import shortestpath.ItemVariations;
+import shortestpath.Util;
+import shortestpath.WorldPointUtil;
 import lombok.extern.slf4j.Slf4j;
 
 /**
@@ -20,7 +20,7 @@ public class Transport {
     public static final int UNDEFINED_ORIGIN = WorldPointUtil.UNDEFINED;
     public static final int UNDEFINED_DESTINATION = WorldPointUtil.UNDEFINED;
     /** A location placeholder different from null to use for permutation transports */
-    private static final int LOCATION_PERMUTATION = WorldPointUtil.packWorldPoint(-1, -1, 1);
+    public static final int LOCATION_PERMUTATION = WorldPointUtil.packWorldPoint(-1, -1, 1);
     private static final String DELIM_SPACE = " ";
     private static final String DELIM_MULTI = ";";
     private static final String DELIM_STATE = "=";
@@ -345,132 +345,5 @@ public class Transport {
             }
         }
         return quests;
-    }
-
-    private static void addTransports(Map<Integer, Set<Transport>> transports, String path, TransportType transportType, int radiusThreshold) {
-        try {
-            String s = new String(Util.readAllBytes(ShortestPathPlugin.class.getResourceAsStream(path)), StandardCharsets.UTF_8);
-            addTransportsFromContents(transports, s, transportType, radiusThreshold);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-
-    private static void addTransports(Map<Integer, Set<Transport>> transports, String path, TransportType transportType) {
-        addTransports(transports, path, transportType, 0);
-    }
-
-    public static void addTransportsFromContents(Map<Integer, Set<Transport>> transports, String contents, TransportType transportType, int radiusThreshold) {
-        final String DELIM_COLUMN = "\t";
-        final String PREFIX_COMMENT = "#";
-
-        Scanner scanner = new Scanner(contents);
-
-        // Header line is the first line in the file and will start with either '#' or '# '
-        String headerLine = scanner.nextLine();
-        headerLine = headerLine.startsWith(PREFIX_COMMENT + " ") ? headerLine.replace(PREFIX_COMMENT + " ", PREFIX_COMMENT) : headerLine;
-        headerLine = headerLine.startsWith(PREFIX_COMMENT) ? headerLine.replace(PREFIX_COMMENT, "") : headerLine;
-        String[] headers = headerLine.split(DELIM_COLUMN);
-
-        Set<Transport> newTransports = new HashSet<>();
-
-        while (scanner.hasNextLine()) {
-            String line = scanner.nextLine();
-
-            if (line.startsWith(PREFIX_COMMENT) || line.isBlank()) {
-                continue;
-            }
-
-            String[] fields = line.split(DELIM_COLUMN);
-            Map<String, String> fieldMap = new HashMap<>();
-            for (int i = 0; i < headers.length; i++) {
-                if (i < fields.length) {
-                    fieldMap.put(headers[i], fields[i]);
-                }
-            }
-
-            Transport transport = new Transport(fieldMap, transportType);
-            newTransports.add(transport);
-
-        }
-        scanner.close();
-
-        /*
-        * A transport with origin A and destination B is one-way and must
-        * be duplicated as origin B and destination A to become two-way.
-        * Example: key-locked doors
-        * 
-        * A transport with origin A and a missing destination is one-way,
-        * but can go from origin A to all destinations with a missing origin.
-        * Example: fairy ring AIQ -> <blank>
-        * 
-        * A transport with a missing origin and destination B is one-way,
-        * but can go from all origins with a missing destination to destination B.
-        * Example: fairy ring <blank> -> AIQ
-        * 
-        * Identical transports from origin A to destination A are skipped, and
-        * non-identical transports from origin A to destination A can be skipped
-        * by specifying a radius threshold to ignore almost identical coordinates.
-        * Example: fairy ring AIQ -> AIQ
-        */
-        Set<Transport> transportOrigins = new HashSet<>();
-        Set<Transport> transportDestinations = new HashSet<>();
-        for (Transport transport : newTransports) {
-            int origin = transport.getOrigin();
-            int destination = transport.getDestination();
-            // Logic to determine ordinary transport vs teleport vs permutation (e.g. fairy ring)
-            if ((origin == UNDEFINED_ORIGIN && destination == UNDEFINED_DESTINATION)
-                || (origin == LOCATION_PERMUTATION && destination == LOCATION_PERMUTATION)) {
-                continue;
-            } else if (origin != LOCATION_PERMUTATION && origin != UNDEFINED_ORIGIN
-                && destination == LOCATION_PERMUTATION) {
-                transportOrigins.add(transport);
-            } else if (origin == LOCATION_PERMUTATION
-                && destination != LOCATION_PERMUTATION && destination != UNDEFINED_DESTINATION) {
-                transportDestinations.add(transport);
-            }
-            if (origin != LOCATION_PERMUTATION
-                && destination != UNDEFINED_DESTINATION && destination != LOCATION_PERMUTATION
-                && (origin == UNDEFINED_ORIGIN || origin != destination)) {
-                transports.computeIfAbsent(origin, k -> new HashSet<>()).add(transport);
-            }
-        }
-        for (Transport origin : transportOrigins) {
-            for (Transport destination : transportDestinations) {
-                // The radius threshold prevents transport permutations from including (almost) same origin and destination
-                if (WorldPointUtil.distanceBetween2D(origin.getOrigin(), destination.getDestination()) > radiusThreshold) {
-                    transports.computeIfAbsent(origin.getOrigin(), k -> new HashSet<>())
-                        .add(new Transport(origin, destination));
-                }
-            }
-        }
-    }
-
-    public static HashMap<Integer, Set<Transport>> loadAllFromResources() {
-        HashMap<Integer, Set<Transport>> transports = new HashMap<>();
-        addTransports(transports, "/transports/transports.tsv", TransportType.TRANSPORT);
-        addTransports(transports, "/transports/agility_shortcuts.tsv", TransportType.AGILITY_SHORTCUT);
-        addTransports(transports, "/transports/boats.tsv", TransportType.BOAT);
-        addTransports(transports, "/transports/canoes.tsv", TransportType.CANOE);
-        addTransports(transports, "/transports/charter_ships.tsv", TransportType.CHARTER_SHIP);
-        addTransports(transports, "/transports/ships.tsv", TransportType.SHIP);
-        addTransports(transports, "/transports/fairy_rings.tsv", TransportType.FAIRY_RING);
-        addTransports(transports, "/transports/gnome_gliders.tsv", TransportType.GNOME_GLIDER, 6);
-        addTransports(transports, "/transports/hot_air_balloons.tsv", TransportType.HOT_AIR_BALLOON, 7);
-        addTransports(transports, "/transports/magic_carpets.tsv", TransportType.MAGIC_CARPET);
-        addTransports(transports, "/transports/magic_mushtrees.tsv", TransportType.MAGIC_MUSHTREE, 5);
-        addTransports(transports, "/transports/minecarts.tsv", TransportType.MINECART);
-        addTransports(transports, "/transports/quetzals.tsv", TransportType.QUETZAL);
-        addTransports(transports, "/transports/seasonal_transports.tsv", TransportType.SEASONAL_TRANSPORTS);
-        addTransports(transports, "/transports/spirit_trees.tsv", TransportType.SPIRIT_TREE, 5);
-        addTransports(transports, "/transports/teleportation_items.tsv", TransportType.TELEPORTATION_ITEM);
-        addTransports(transports, "/transports/teleportation_boxes.tsv", TransportType.TELEPORTATION_BOX);
-        addTransports(transports, "/transports/teleportation_levers.tsv", TransportType.TELEPORTATION_LEVER);
-        addTransports(transports, "/transports/teleportation_minigames.tsv", TransportType.TELEPORTATION_MINIGAME);
-        addTransports(transports, "/transports/teleportation_portals.tsv", TransportType.TELEPORTATION_PORTAL);
-        addTransports(transports, "/transports/teleportation_spells.tsv", TransportType.TELEPORTATION_SPELL);
-        addTransports(transports, "/transports/wilderness_obelisks.tsv", TransportType.WILDERNESS_OBELISK);
-        return transports;
     }
 }
