+++ src/main/java/shortestpath/transport/TransportItems.java
@@ -1,4 +1,4 @@
-package shortestpath;
+package shortestpath.transport;
 
 import java.util.Arrays;
