+++ src/main/java/shortestpath/transport/TransportVarCheck.java
@@ -1,4 +1,4 @@
-package shortestpath;
+package shortestpath.transport;
 
 import lombok.Getter;
