+++ src/main/java/shortestpath/transport/TransportVarPlayer.java
@@ -1,16 +1,16 @@
-package shortestpath;
+package shortestpath.transport;
 
 import java.util.Map;
 
 import lombok.Getter;
 
-public class TransportVarbit {
+public class TransportVarPlayer {
     @Getter
     private final int id;
     private final int value;
     private final TransportVarCheck check;
 
-    public TransportVarbit(int id, int value, TransportVarCheck check) {
+    public TransportVarPlayer(int id, int value, TransportVarCheck check) {
         this.id = id;
         this.value = value;
         this.check = check;
