+++ src/test/java/shortestpath/pathfinder/PathfinderTest.java
@@ -22,10 +22,12 @@ import shortestpath.ItemVariations;
 import shortestpath.ShortestPathConfig;
 import shortestpath.ShortestPathPlugin;
 import shortestpath.TeleportationItem;
-import shortestpath.Transport;
-import shortestpath.TransportItems;
-import shortestpath.TransportType;
 import shortestpath.WorldPointUtil;
+import shortestpath.transport.Transport;
+import shortestpath.transport.TransportItems;
+import shortestpath.transport.TransportLoader;
+import shortestpath.transport.TransportType;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
@@ -35,7 +37,7 @@ import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class PathfinderTest {
-    private static final Map<Integer, Set<Transport>> transports = Transport.loadAllFromResources();
+    private static final Map<Integer, Set<Transport>> transports = TransportLoader.loadAllFromResources();
 
     private PathfinderConfig pathfinderConfig;
